Refactor NCMenu
Makes it easy for any ViewController to display a menu, reduce code duplication
