+++ iOSClient/Menu/NCLoginWeb+Menu.swift
@@ -28,7 +28,6 @@ extension NCLoginWeb {
 
     func toggleMenu() {
         
-        let menuViewController = UIStoryboard.init(name: "NCMenu", bundle: nil).instantiateInitialViewController() as! NCMenu
         var actions = [NCMenuAction]()
         
         let accounts = NCManageDatabase.shared.getAllAccount()
@@ -82,16 +81,8 @@ extension NCLoginWeb {
                 }
             )
         )
-       
-        menuViewController.actions = actions
 
-        let menuPanelController = NCMenuPanelController()
-        menuPanelController.parentPresenter = self
-        menuPanelController.delegate = menuViewController
-        menuPanelController.set(contentViewController: menuViewController)
-        menuPanelController.track(scrollView: menuViewController.tableView)
-
-        self.present(menuPanelController, animated: true, completion: nil)
+        presentMenu(with: actions)
     }
 }
